Changes
I'm stuck in the ajax function for generating the students list.
I mean I can generate everything, even the template list but the issue
is that the $.get is inside the each function, meaning I have to get the
ajax data out of the $.get.
I have tried different coding styles since 1990 to no avail.