More classroom functions
Added a button and its functions for add more Students in classroom.
Added remove students from classroom button and its functions.
Made the action button for esomoModal (the one with the list of students
to be added/removed from the classroom) to be dynamic depending on what
the action is meant to be.

I will pull changes, see if there are any bugs then we get finished with
classrooms